// embedded-components/src/api/generated/smbdo.schemas.ts
export interface AttestationRemoval {
  documentId: DocumentId;
}

/**
 * Update a party inline within the clients API by supplying its ID.
 */
export interface UpdatePartyRequestInline {
  access?: AccessList;
  email?: Email;
  id?: PartyId;
  individualDetails?: IndividualDetails;
  organizationDetails?: OrganizationDetails;
  /**
   * @minItems 1
   * @maxItems 10
   */
  roles?: Role[];
}

export type UpdateClientRequestSmbdoAddPartiesItem = CreatePartyRequestInline &
  UpdatePartyRequestInline;

export interface UpdateClientRequestSmbdo {
  /**
   * @minItems 1
   * @maxItems 10
   */
  addAttestations?: Attestation[];
  /**
   * Update a party by specifying its ID in the object. Create a party by setting `parentPartyId` to the client's root party. `partyType` and `roles` are required when creating a new party.

   * @minLength 1
   * @maxLength 10
   * @minItems 0
   */
  addParties?: UpdateClientRequestSmbdoAddPartiesItem[];
  addProducts?: ClientProductList;
  consumerDevice?: ConsumerDevice;
  /**
   * @minItems 0
   * @maxItems 200
   */
  questionResponses?: ClientQuestionResponse[];
  /**
   * @minItems 1
   * @maxItems 10
   */
  removeAttestations?: AttestationRemoval[];
}
// ...
export type ResponseValueList = string[];

/**
 * The client's response to a question.
 */
export interface ClientQuestionResponse {
  questionId?: QuestionId;
  values?: ResponseValueList;
}
// ...
export type OrganizationDetailsRequired = OrganizationDetails;

/**
 * Create a party within the clients payload.
 */
export interface CreatePartyRequestInline {
  access?: AccessList;
  email?: Email;
  externalId?: ExternalId;
  individualDetails?: IndividualDetailsRequired;
  organizationDetails?: OrganizationDetailsRequired;
  parentPartyId?: ParentPartyId;
  partyType?: PartyTypeSmbdo;
  /**
   * @minItems 1
   * @maxItems 10
   */
  roles?: Role[];
}

/**
 * Create a party within the clients payload.
 */
export type CreatePartyRequestInlineRequired = CreatePartyRequestInline;

export interface CreateClientRequestSmbdo {
  /**
   * @minItems 1
   * @maxItems 10
   */
  attestations?: Attestation[];
  consumerDevice?: ConsumerDevice;
  /**
   * @minItems 1
   * @maxItems 10
   */
  parties?: CreatePartyRequestInlineRequired[];
  partyId?: PartyId;
  products: ClientProductList;
}
// ...
export type IndividualDetailsRequired = IndividualDetails;
// ...
export type PartyTypeSmbdo =
  (typeof PartyTypeSmbdo)[keyof typeof PartyTypeSmbdo];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PartyTypeSmbdo = {
  INDIVIDUAL: 'INDIVIDUAL',
  ORGANIZATION: 'ORGANIZATION',
} as const;

/**
 * The parent party ID.

 * @minLength 10
 * @maxLength 10
 */
export type ParentPartyId = string;

/**
 * Id in external system.
 * @minLength 1
 * @maxLength 20
 */
export type ExternalId = string;

/**
 * Email of the party.
 * @minLength 0
 * @maxLength 320
 */
export type Email = string;

/**
 * List of user access descriptors.
 * @deprecated
 * @minItems 0
 * @maxItems 10
 */
export type AccessList = string[];

/**
 * Consumer device information provided by the merchant. It is only returned when present in the create or update client requests.
 */
export interface ConsumerDevice {
  /** IPv4 address of the consumer device. */
  ipAddress?: string;
  /**
   * The unique session identifier of the device that was created by the client. It can be up to 128 characters long and can contain only the following characters: uppercase and lowercase Roman letters, digits, underscore characters, and hyphens (a?z, A?Z, 0?9, _, -). The session ID should contain at least 16 bytes of randomly generated data.

   * @pattern ^[a-zA-Z0-9_-]+$
   */
  sessionId?: string;
}

/**
 * The attesting of presented legal statements. Either "attester" or "attesterFullName" is required.
 */
export interface Attestation {
  attestationTime: string;
  attester?: Attester;
  /** The full name of an individual. */
  attesterFullName?: string;
  documentId: DocumentId;
  /** IPv4 Address. */
  ipAddress: string;
}
// ...
export interface Attester {
  designation: AttesterDesignation;
  firstName: AttesterFirstName;
  lastName: AttesterLastName;
  middleName?: AttesterMiddleName;
}
// ...
export type Role = (typeof Role)[keyof typeof Role];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const Role = {
  AUTHORIZED_USER: 'AUTHORIZED_USER',
  BENEFICIAL_OWNER: 'BENEFICIAL_OWNER',
  CLIENT: 'CLIENT',
  CONTROLLER: 'CONTROLLER',
  DECISION_MAKER: 'DECISION_MAKER',
  PRIMARY_CONTACT: 'PRIMARY_CONTACT',
  DIRECTOR: 'DIRECTOR',
} as const;

/**
 * A list of party roles.
 * @minItems 0
 * @maxItems 10
 */
export type PartyRoleList = Role[];

/**
 * The ID uniquely identifying a question.
 * @maxLength 10
 */
export type QuestionId = string;

/**
 * A list of question IDs.
 * @minItems 0
 * @maxItems 200
 */
export type QuestionIdList = QuestionId[];

/**
 * A list of document request IDs.
 * @minItems 0
 * @maxItems 10
 */
export type DocumentRequestIdList = string[];

/**
 * The unique id generated by the system for the uploaded document,  which can be used for future retrieval.

 */
export type DocumentId = string;
// ...
export type ClientProduct = (typeof ClientProduct)[keyof typeof ClientProduct];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ClientProduct = {
  MERCHANT_SERVICES: 'MERCHANT_SERVICES',
  EMBEDDED_PAYMENTS: 'EMBEDDED_PAYMENTS',
} as const;

/**
 * A list of product selections.
 * @minItems 1
 * @maxItems 50
 */
export type ClientProductList = ClientProduct[];

/**
 * Organization/Individual party id, that uniquely identifies the party.

 * @maxLength 10
 * @pattern ^[0-9]{10}$
 */
export type PartyId = string;